genie-slack-app/requirements.txt
slack-bolt==1.18.0
slack-sdk==3.27.1
databricks-sdk==0.23.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

